Refactor(jira): Create some packages and move classes into it.
