+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/config/IssueConfig.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.issuetracker;
+package com.synopsys.integration.alert.issuetracker.config;
 
 import java.util.Optional;
