+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/jira/common/util/JiraTransitionHandler.java
@@ -29,9 +29,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.synopsys.integration.alert.common.enumeration.ItemOperation;
-import com.synopsys.integration.alert.issuetracker.IssueConfig;
-import com.synopsys.integration.alert.issuetracker.IssueMissingTransitionException;
 import com.synopsys.integration.alert.issuetracker.TransitionValidator;
+import com.synopsys.integration.alert.issuetracker.config.IssueConfig;
+import com.synopsys.integration.alert.issuetracker.exception.IssueMissingTransitionException;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.jira.common.model.components.IdComponent;
 import com.synopsys.integration.jira.common.model.components.StatusCategory;
